+++ src/mocks/index.ts
@@ -1,2 +1,2 @@
 export * from "./posts";
-export * from "./DataDemo";
+export * from "./users";
