+++ src/shared/components/FollowButton.tsx
@@ -1,0 +1,21 @@
+import { Button } from "@chakra-ui/react";
+
+interface Props {
+  title: string;
+  onOpen?: () => void;
+}
+export const FollowButton = ({ title, onOpen }: Props) => {
+  return (
+    <Button
+      onClick={onOpen}
+      backgroundColor={"#FF5050"}
+      size={"xs"}
+      paddingX={"15px"}
+      textColor={"#E0E0E0"}
+      _hover={{ bg: "#FF5050" }}
+      rounded={"2xl"}
+    >
+      {title}
+    </Button>
+  );
+};
