+++ src/shared/components/index.ts
@@ -1,17 +1,15 @@
-export * from "./RightSidebar";
-export * from "./LeftSidebar";
-export * from "./LeftSidebarLinks";
-export * from "./NavLink";
-export * from "./MainContainer";
-export * from "./Sidebar";
-export * from "./CommentsModal";
-export * from "../../auth/components/LogIn";
-export * from "../../auth/components/RecoveryPassword";
-export * from "../../auth/components/CreateAccount";
-export * from "../../auth/components/RightSidebarNoLogin";
-export * from "./SettingForm";
-export * from "./InputSetting";
 export * from "./ButtonComponent";
 export * from "./CommentCard";
-export * from "./FollowBoton";
+export * from "./CommentsModal";
+export * from "./FollowButton";
+export * from "./InputComponent";
+export * from "./InputSetting";
+export * from "./LeftSidebar";
+export * from "./LeftSidebarLinks";
+export * from "./MainContainer";
+export * from "./ModalContainForm";
+export * from "./NavLink";
+export * from "./RightSidebar";
+export * from "./SettingForm";
+export * from "./Sidebar";
 export * from "./UserCard";
